requirements.txt
Adafruit-Blinka==7.1.1
adafruit-circuitpython-dht==3.7.1
Adafruit-PlatformDetect==3.21.0
Adafruit-PureIO==1.1.9
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
bidict==0.21.4
CacheControl==0.12.10
cachy==0.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cleo==0.8.1
click==8.0.3
clikit==0.6.2
colorzero==2.0
configparser==5.2.0
crashtest==0.3.1
cryptography==36.0.2
Deprecated==1.2.13
distlib==0.3.4
fastapi==0.75.0
filelock==3.6.0
frozenlist==1.3.0
gpiozero==1.6.2
h11==0.13.0
html5lib==1.1
httptools==0.3.0
idna==3.3
importlib-metadata==4.11.3
jeepney==0.7.1
keyring==23.5.0
lockfile==0.12.2
msgpack==1.0.3
multidict==6.0.2
packaging==20.9
paho-mqtt==1.6.1
pastel==0.2.1
pexpect==4.8.0
pigpio==1.78
pkginfo==1.8.2
platformdirs==2.5.1
poetry-core==1.0.8
ptyprocess==0.7.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pyftdi==0.54.0
pylev==1.4.0
pyparsing==3.0.7
pyserial==3.5
python-dotenv==0.19.2
python-engineio==4.3.1
python-socketio==5.5.1
pyusb==1.2.1
PyYAML==6.0
redis==4.2.2
requests==2.27.1
requests-toolbelt==0.9.1
SecretStorage==3.3.1
shellingham==1.4.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
tomlkit==0.10.0
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.1
uvloop==0.16.0
virtualenv==20.13.3
watchgod==0.7
webencodings==0.5.1
websockets==10.1
wrapt==1.14.0
yarl==1.7.2
zipp==3.7.0
